refactor: clean-up app layout
