+++ src/components/layout/ScreenBoundingBox.tsx
@@ -1,0 +1,16 @@
+import { Box } from "@mui/material";
+import { FC, PropsWithChildren } from "react";
+
+const ScreenBoundingBox: FC<PropsWithChildren> = ({ children }) => {
+    return (
+        <Box
+            style={{
+                display: "flex"
+            }}
+        >
+            {children}
+        </Box>
+    );
+};
+
+export default ScreenBoundingBox;
